backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["arkadiy <cotnikoarkady@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.4"
langserve = {extras = ["all"], version = "^0.3.0"}
langchain = "^0.3.5"
langchain-openai = "^0.2.4"
pydantic-settings = "^2.6.0"
toml = "^0.10.2"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.7"
faiss-cpu = "^1.9.0"
langchain-mistralai = "^0.2.4"
chromadb = "^0.5.23"
tqdm = "^4.67.1"
langchain-chroma = "^0.1.4"
langchain-huggingface = "^0.1.2"
nltk = "^3.9.1"
rank-bm25 = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
